+++ CHANGELOG.md
@@ -20,6 +20,8 @@ This update accompanies the release of the [Telegram Bot API 6.6](https://core.t
    using the method `\TgBotLib\Bot > getMyShortDescription()` see [getMyShortDescription](https://core.telegram.org/bots/api#getmyshortdescription) for more information.
  * Added method `\TgBotLib\Bot > sendSticker()` to send a sticker to a chat. See [sendSticker](https://core.telegram.org/bots/api#sendsticker) for more information.
  * Added method `\TgBotLib\Bot > createNewStickerSet()` to create a new sticker set owned by a user. See [createNewStickerSet](https://core.telegram.org/bots/api#createnewstickerset) for more information.
+ * Added the field _needs_repainting_ to the [Sticker](https://core.telegram.org/bots/api#sticker) class
+   which can be obtained via `TgBotLib\Objects\Telegram > Sticker > needsRepainting()`
 
 ### Changed
  * Removed unused `__destruct()` method from `\TgBotLib\Bot`